Refactoring compatibility.py and prep for testing
