+++ pc-advisor/__init__.py
@@ -1,0 +1,10 @@
+"""
+PC Build Advisor
+----------------
+A tool for recommending and validating PC builds using a local parts dataset
+and GPT-4o for final recommendations.
+
+Note on pricing:
+    Prices in the dataset were scraped from PCPartPicker at a point in time
+    and may be outdated. GPT-4o is instructed to acknowledge this.
+"""
