+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+testpaths = tests
+addopts = --cov=pc_advisor --cov-report=term-missing
